Refactor code to add a tree output with changes in the publish endpoint.
The code has been refactored to print out a tree that includes a report of package changes when executing a publish operation. This was achieved by expanding the functionality of various structs throughout the different modules involved in the publishing sequence, such as BumpReport and ExtendedBumpReport. Now, more detailed information such as package names, old and new versions, and files affected by the operation are provided.