Update CommandFormer with expanded handler functionality
The update introduces changes to ‘CommandFormer’ in the ‘command.rs’ file of the 'ca grammar' module. It introduces ability to set command routine types, such as those with arguments, properties, a context, or combinations thereof. The type `Routine` has been modified and a panic message is included for instances when a handler function for the command is missing. The changes also incorporate expanded use of generics and trait boundaries.