Update test method in wca_hello_test
The main action of this commit is updating the test method in wca_hello_test. An import statement was added at the start of the file to include Args from module wca. Additionally, the routine method was updated to use the '.end()' method instead of '.perform()'. These changes were necessary to properly test the functionalities of the wca module.