Change visibility to exposed for several modules
The scope of several modules has been changed from 'prelude' to 'exposed'. This results in the wider availability of these modules within the project. The modules affected include Args, Props, Type, TryCast, Context, and CommandsAggregator. The corresponding changes in use statements have been made in the associated code and documentation.