Add `exclude` and `has_changes` methods to `DiffReport`
The `DiffReport` struct is now cloneable, and received two new methods: `exclude` and `has_changes`. The former allows for exclusion of specified items from a report, while the latter checks if there are changes within the `DiffItems`. The commit also modifies the `crate_diff` function to compare two crate archives more accurately, as well as incorporating the `IGNORE_LIST` in the `publish_diff` and `package` modules to exclude certain files from the differential report.