Refactor help generator and implement table formatting tool
HelpGeneratorArgs has been renamed to HelpGeneratorOptions to better reflect its purpose. The help content generator now makes use of a new table formatting tool which simplifies the generation and formatting of tables. This tool exists in the newly created 'tool' module which might encompass more general-purpose tools in the future.