Refactor Parser to handle command line arguments
This commit refactors the Parser module of the wca library, enabling it to parse command-line arguments as a vector of Strings. This optimises functionality by enabling commands to be parsed directly from arguments passed to a program's main function. Additional improvements include further structuring the parse_command function and removal of unnecessary dependencies on the nom crate and methods from Parser.