Refactor publish action to use a publication plan
The publish action has been modified to be handle via a publication plan. This adds an intermediary step where the publication steps are defined first in a plan before execution. This change also removed the 'wanted_to_publish' attribute from the PublishReport struct. Displaying as a tree is now also based on roots within the plan.